requirements.txt
pynacl
simplejson
pytest
